requirements.txt
Django
django-filter
djangorestframework
rest_framework.authtoken
python-decouple